# specs/003-user-auth/contracts/auth-api.openapi.yaml
openapi: 3.0.3
info:
  title: Collab Todo - Authentication & Authorization API
  description: |
    API specification for user authentication and authorization features in the collaborative to-do application.
    
    ## Authentication
    - Uses JWT tokens stored in HTTP-only cookies
    - Session expiration: 24 hours
    - Cookie name: `auth_token`
    
    ## Authorization Hierarchy
    1. **Owner** (authenticated user who created list): Full control (read, write, delete)
    2. **Edit Link** (shareId): Read and write access (no delete)
    3. **View Link** (viewId): Read-only access
    
    ## Rate Limiting
    - Signup: 3 attempts per IP per hour
    - Signin: 5 attempts per email per 15 minutes
    
    ## Security
    - All endpoints use HTTPS in production
    - Passwords hashed with bcrypt (cost factor 12)
    - Sensitive endpoints require authentication
  version: 1.0.0
  contact:
    name: Collab Todo API Support
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.collabtodo.app
    description: Production server
...
paths:
  /api/auth/signup:
    post:
      tags:
        - Authentication
      summary: Create a new user account
      description: |
        Register a new user with email and password. Returns user object and sets authentication cookie.
        
        **Rate Limit**: 3 attempts per IP per hour
        
        **Validation**:
        - Email must be valid format (RFC 5322)
        - Email must be unique (case-insensitive)
        - Password must be minimum 8 characters
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  minLength: 5
                  maxLength: 254
                  example: alice@example.com
                  description: User's email address (stored lowercase)
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: SecurePass123
                  description: User's password (will be hashed with bcrypt)
      responses:
        '201':
          description: User created successfully, authentication cookie set
          headers:
            Set-Cookie:
              description: HTTP-only cookie with JWT token
              schema:
                type: string
                example: auth_token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...; HttpOnly; Secure; SameSite=Strict; Path=/; Max-Age=86400
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid request (validation error)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidEmail:
                  summary: Invalid email format
                  value:
                    error: Invalid email format
                invalidPassword:
                  summary: Password too short
                  value:
                    error: Password must be at least 8 characters
                duplicateEmail:
                  summary: Email already registered
                  value:
                    error: Email already registered
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Seconds until rate limit resets
              schema:
                type: integer
                example: 3600
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Too many signup attempts. Please try again later.
  /api/auth/signin:
    post:
      tags:
        - Authentication
      summary: Sign in to existing account
      description: |
        Authenticate user with email and password. Returns user object and sets authentication cookie.
        
        **Rate Limit**: 5 attempts per email per 15 minutes
        
        **Security Notes**:
        - Returns same error message for invalid email or password (prevents email enumeration)
        - Uses constant-time password comparison (prevents timing attacks)
      operationId: signin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: alice@example.com
                password:
                  type: string
                  format: password
                  example: SecurePass123
      responses:
        '200':
          description: Signin successful, authentication cookie set
          headers:
            Set-Cookie:
              description: HTTP-only cookie with JWT token
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Invalid credentials
        '429':
          description: Rate limit exceeded
          headers:
            Retry-After:
              description: Seconds until rate limit resets
              schema:
                type: integer
                example: 900
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Too many signin attempts. Please try again later.
...
  /api/lists/{shareId}/items/{itemId}:
    patch:
      tags:
        - Lists
      summary: Update item (complete/uncomplete, edit text)
      description: |
        **Authorization**: Owner OR valid edit link (shareId)
      operationId: updateItem
      security:
        - cookieAuth: []
        - {}
      parameters:
        - name: shareId
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
                  maxLength: 500
                completed:
                  type: boolean
      responses:
        '200':
          description: Item updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  item:
                    $ref: '#/components/schemas/TodoItem'
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Item or list not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Lists
      summary: Delete item
      description: |
        **Authorization**: Owner OR valid edit link (shareId)
      operationId: deleteItem
      security:
        - cookieAuth: []
        - {}
      parameters:
        - name: shareId
          in: path
          required: true
          schema:
            type: string
        - name: itemId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Item deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Item deleted successfully
        '403':
          description: Insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Item or list not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: JWT token stored in HTTP-only cookie
  schemas:
    User:
      type: object
      required:
        - _id
        - email
        - createdAt
        - updatedAt
      properties:
        _id:
          type: string
          format: objectId
          example: 65a1b2c3d4e5f6a7b8c9d0e1
          description: Unique user identifier (MongoDB ObjectId as string)
        email:
          type: string
          format: email
          example: alice@example.com
          description: User's email address
        createdAt:
          type: string
          format: date-time
          example: '2026-01-06T10:30:00.000Z'
          description: Account creation timestamp (ISO 8601)
        updatedAt:
          type: string
          format: date-time
          example: '2026-01-06T10:30:00.000Z'
          description: Last account update timestamp (ISO 8601)
...
    TodoItem:
      type: object
      required:
        - _id
        - text
        - completed
        - order
        - createdAt
        - updatedAt
      properties:
        _id:
          type: string
          format: objectId
          example: 65a1b2c3d4e5f6a7b8c9d0e3
        text:
          type: string
          maxLength: 500
          example: Write API documentation
        completed:
          type: boolean
          example: false
        order:
          type: integer
          example: 0
          description: Display order (for sorting)
        createdAt:
          type: string
          format: date-time
          example: '2026-01-06T11:05:00.000Z'
        updatedAt:
          type: string
          format: date-time
          example: '2026-01-06T11:05:00.000Z'
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Invalid email format
          description: Human-readable error message